requirements.txt
beautifultable==1.1.0
cryptography==38.0.1
inquirer==3.1.3
pwinput==1.0.3
pyOpenSSL==23.2.0
termcolor==2.3.0